fix: align MCP server with existing flair-mcp package
- remove duplicate packages/mcp-server
- use Basic auth from FLAIR_ADMIN_USER/FLAIR_ADMIN_PASSWORD for memory_search and memory_store
- parse /SemanticSearch response as { results: [...] }
- default FLAIR_URL to http://localhost:19926
- update flair-mcp README defaults to 19926
