--- packages/mcp-server/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@tpsdev-ai/mcp-server",
-  "version": "0.1.0",
-  "description": "Basic MCP wrapper for Flair HTTP endpoints",
-  "type": "module",
-  "main": "dist/index.js",
-  "bin": {
-    "flair-mcp-server": "dist/index.js"
-  },
-  "scripts": {
-    "build": "tsc -p tsconfig.json"
-  },
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.12.1",
-    "zod": "^4.3.6"
-  },
-  "devDependencies": {
-    "@types/node": "^24.11.0",
-    "typescript": "5.9.3"
-  }
-}
